% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{calc_bmi}
\alias{calc_bmi}
\title{Title- Calculate bmi.}
\usage{
calc_bmi(weight = 0, height = 0)
}
\arguments{
\item{weight-}{A numeric or a float variable}

\item{height-}{A numeric or a float variable}
}
\value{
bmi- A numeric or a float variable
}
\description{
Title- Calculate bmi.
}
\examples{
calc_bmi(57, 172) or
medicaldata::smartpill |>
   as_tibble() |>
   select(Weight, Height) |>
   mutate(BMI = calc_bmi(Weight, Height))
}
