% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{calculate_calories2}
\alias{calculate_calories2}
\title{Title: Calculate calories from macronutrients.}
\usage{
calculate_calories2(carb_grams = 0, protein_grams = 0, fat_grams = 0)
}
\arguments{
\item{carb_grams-}{a numeric or float variable}

\item{protein_grams-}{a numeric or float variable}

\item{fat_grams-}{a numeric or float variable}
}
\value{
a numeric variable or a float
}
\description{
Title: Calculate calories from macronutrients.
}
\examples{
calculate_calories(10, 5, 15)
calculate_calories(10, 5, 0)
}
