% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{calculate_isoniazid_dosage2}
\alias{calculate_isoniazid_dosage2}
\title{Title- Calculate isoniazide dosage.-}
\usage{
calculate_isoniazid_dosage2(weight)

calculate_isoniazid_dosage2(weight)
}
\arguments{
\item{weight-}{A numeric, a float, a vector or or a numeric variable in a DataFrame}
}
\value{
- A dosage, which must be numeric

- A dosage, which must be numeric
}
\description{
Title- Calculate isoniazide dosage.-

Title- Calculate isoniazide dosage.-
}
\examples{
calculate_isoniazid_dosage2(c(42, 30, 70)) or

medicaldata::smartpill |>
as_tibble() |>
 select(Weight) |>
 mutate(dosage = calculate_isoniazid_dosage2(Weight))
calculate_isoniazid_dosage2(c(42, 30, 70)) or

medicaldata::smartpill |>
as_tibble() |>
 select(Weight) |>
 mutate(dosage = calculate_isoniazid_dosage2(Weight))
}
