% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{celsius_convert}
\alias{celsius_convert}
\title{Title- Converting celsius scale to other common scale.}
\usage{
celsius_convert(celsius, convert_to)
}
\arguments{
\item{celsius-}{A numeric or a float variable}

\item{convert_to-}{The scale to convert to, e.g,fahrenheit, kelvin, centigrade. It must be a string.}
}
\value{
- A numeric or a float
}
\description{
Title- Converting celsius scale to other common scale.
}
\examples{
celsius_convert(36.6, "fahrenheit")
}
