% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Title- Converting Celsius to Fahrenheit.-}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius-}{A numeric or float variable}
}
\value{
a number
}
\description{
Title- Converting Celsius to Fahrenheit.-
}
\examples{
celsius_to_fahrenheit(37.5)
}
