% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{classify_temp}
\alias{classify_temp}
\title{Title- Classify body temperatue of humans.-}
\usage{
classify_temp(temp)
}
\arguments{
\item{temp-}{A numeric or a float variable}
}
\value{
- a string
}
\description{
Title- Classify body temperatue of humans.-
}
\examples{
classify_temp(36.9)
}
