% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Title- Converting Fahrenheit to Celsius.}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit-}{A numeric variable}
}
\value{
a number
}
\description{
Title- Converting Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(97)
}
