% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{pounds_to_kg}
\alias{pounds_to_kg}
\title{Title- # Converting pounds to kilograms -}
\usage{
pounds_to_kg(pounds)
}
\arguments{
\item{pounds-}{A numeric value-}
}
\value{
a number or a float-
}
\description{
Title- # Converting pounds to kilograms -
}
\examples{
pounds_to_kg(75)-
}
