% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{years_to_months}
\alias{years_to_months}
\title{Title- Transforms age in years to age in months}
\usage{
years_to_months(years)
}
\arguments{
\item{years-}{A numeric variable}
}
\value{
a number
}
\description{
Title- Transforms age in years to age in months
}
\examples{
years_to_months(3)
}
